ingest: start Snakefile with fetch_sequences rules
Currently only fetches sequences from GenBank, but other sources can be
defined in the config. If adding other sources, add a new rule upstream
of rule `fetch_all_sequences` in `fetch_sequences.smk` to create the
file `data/{source}.ndjson` or the file must exist as a static file in
the repo.

All sequences are concatenated in a single NDJSON file that can then
be provided to downstream transform rules.