ingest: start Snakemake rule for transform
Adds script `bin/ndjson-to-tsv-and-fasta` that parses NDJSON records
from stdin to a metadata TSV and a FASTA file. Drops any records that
do not have a sequence ID or sequence to ensure the produced FASTA file
does not contain empty records.

Currently the transform rule only transfroms the NDJSON records to
the metadata TSV and FASTA files.