ingest/transform: add `transform-strain-names`
Checks the strain name in the `strain` field of the NDJSON records
match the provided regex pattern. Currently just using `'^.+$'` to
accept any strain name because we do not have a clear standard yet
for strain names. Once we have a standard pattern, this regex can just
be updated in the `config.yaml`.

We are currently using the `accession` field as a backup field for
strain names since there are many records from GenBank that do not have
a strain name.