ingest: add rules for uploading files
Currently only includes rule to upload files to AWS S3, but additional
upload targets can be added by adding new rules to `upload.smk`
and including the new targets in `config.upload`. Currently all upload
params as defined in the `config/optional.yaml` file to separate it from
the main ingest pipeline params.

The use of wildcards within `_get_all_targets` for creating the targets
for upload allows each file upload to be created as a separate job.
This allows each file to uploaded as soon as they are generated,
and multiple files can be uploaded simultaneously.