ingest/slack_notifications: add `onstart` and `onerror`
`onstart` sends a Slack notification of the ingest job starting.
Provides links to the GitHub action AND AWS Batch job if their
respective IDs are available as environment variables.

`onerror` sends a Slack notification if the ingest job fails. Provides
the link to the AWS Batch job OR the GitHub action if the ID is
available as an environment variable.

Added to the `slack_notifications` module because these will only need
to run if the Slack notifications are toggled on. However, they will be
ignored if there are other `onstart`/`onerror` handlers further down
the Snakefile.