ingest/transform-date-fields: make `format_date` more robust
Previous implementation of masking incomplete dates was dependent on
the date string using '-' to separate date fields. Providing date
strings with any other separator would result in a masked date.

This commit updates `format_date` to check directives within the date
format to verify if year, month, and day are included in the date format.

If date format does not contain any directives for year, then return a
completely masked date because a date without the year is completely
useless to us.

If the date format does not contain any directives for month,
then mask both month and day fields with 'XX' because just year and day
is completely useless to us.