ingest/transform-date-fields: keep `expected-date-formats` order
There are potential date strings that can match multiple formats.
For example, '2020-01' can match both '%Y-%m' and '%Y-%d'. I'm not sure
how we can robustly detect these clashing formats. So the next best
thing is to respect the order in which the `expected-date-formats`
are provided and use the first matching format in the list.