ingest/apply-geolocation-rules: update general rules logic
Previously, we would backfill `rule_traversal` with wildcards ("*")
if we could not find a matching rule for a particular `raw_geolocation`.
This would NOT work for cases where there are partial rule matches AND
wildcard rules that match. I realized the flaw in this logic
while responding to @victorlin's post-merge review:
https://github.com/nextstrain/monkeypox/pull/41#discussion_r896253672

This commit updates the logic for when there are no matching rules. The
`rule_traversal` is reset to the last index that is currently not a
wildcard rule, and then change this value to a wildcard. This allows the
recursive function to try different iterations of `rule_traversal` with
different combinations of raw values and wildcards.