Require augur version 16.0.0
This is necessary for the temporal colour scale introduced in the
preceding commit.

For searchibility i'll past the error message
that you would get when running with <16.0.0 (which shouldn't be
reachable as the snakemake pipeline should exit up front):

ERROR: 'temporal' is not one of ['continuous', 'ordinal', 'categorical', 'boolean']
Validation of config/auspice_config_mpxv.json failed.