Subsample MPXV clades
As more and more B.1 viruses are sequenced, the MPXV global diversity tree is becoming increasingly lopsided and not very descriptive of broader diversity.

This commit updates filter to subsample based on "clade lineage" so that every clade / lineage category is subsampled to. It sets sequences per group to 1000 for hMPXV so that no subsampling occurs and sets sequences per group to 100 for MPXV, which effectively just subsamples B.1 from 289 genomes to 100.