ingest: Update schedule to run daily at 4pm UTC
During the internal Nextstrain meeting on July 21, 2022, the team
decided to reduce the monkeypox/ingest pipeline to run only once a day
to reduce bombardment of Slack messages.

Through Slack discussion¹, we decided to run it daily at 12pm ET since
the curation by NCBI is done on the East Coast.

¹ https://bedfordlab.slack.com/archives/C03G8HQEV18/p1658860515668579