ingest/upload: Wait for Slack notification outputs
In cases where we use Slack notification scripts to do diffs with
S3 files, we want to run the notification rules _before_ we upload the
latest files to S3. Prior to this change, there was a race condition
between the `upload_to_s3` rule and the `notify_on_genbank_record_change`
rule which could lead to the diff showing "No new records".

This commit adds an input function for `upload_to_s3` that keeps track
of the flag files from `send_notifications.smk` that need to be included
as inputs to ensure the rules run in the correct order.