Add deploy rule for automated builds
Uses `nextstrain remote upload` command to deploy builds to a specified
`deploy_url` provided in the config. Purposely uses the build name from
the config instead of the wildcard build name to allow the rule to be
a target rule without providing explicit output files.

The default `deploy_url` is added to an optional config file
`config/nextstrain_automation.yaml` that is intended for internal use
by the Nextstrain team.