ingest/notify-slack: support threaded messages
The `--output` option allows us to save the curl output in a file.
This will allow us to parse the response for the Slack thread timestamp
required to post threaded messages.

The `--thread-ts` option allows us to provide a timestamp for the
Slack message to be sent as a threaded message.

The `--broadcast` option allows us to send a threaded Slack message that
are also broadcasted to the channel for high priority messages such
as error messages.

This will make the Slack notifications for automated builds much cleaner
where each build will have their own start message with threaded
replies for further updates.