Add `onstart` handler for automated builds
Sends a Slack notification that indicates the start of the build
pipeline. Includes the AWS Batch and GitHub Action links if their ids
are provided via environment variables `AWS_BATCH_JOB_ID` and
`GITHUB_RUN_ID`.

Saves the onstart Slack message timestamp to a file so that it can be
accessed by later notification jobs to thread their responses.