ingest: trigger rebuilds for new data only
We do not want to waste time and resources to run rebuilds when there's
no new data from ingest. This commit checks the output from the
uploads of the metadata TSV and sequences FASTA files to verify they
have been uploaded to S3. If the output includes the message
"files are identical" then we know they have not been updated and skip
the trigger for rebuilds.