Snakefile: Set AUGUR_RECURSION_LIMIT=10000
This change is a response to the recursion error that a user ran into
when running the workflow with a large number of sequences during
Nextstrain office hours on 2023-04-13.

Set the recursion limit globally instead of asking the user to remember
to set the environment variable or edit their own Snakefile.