fix: snakemake warning on geolocation filepath
This path when used as an input causes Snakemake to produce a warning
(emitted at the beginning of the logs) and mentioned here:
https://github.com/nextstrain/dengue/pull/6#discussion_r1149887926

> Relative file path './source-data/geolocation-rules.tsv' starts with './'.
This is redundant and strongly discouraged. It can also lead to inconsistent
results of the file-matching approach used by Snakemake. You can simply omit
the './' for relative file paths.

This commit removes the `./` from the filepath to avoid the warning.