[wrangle metadata] Uses csvtk instead of a separate script
Uses csvtk instead of a separate wrangle script as discussed in this thread:

https://github.com/nextstrain/zika/pull/25#discussion_r1117734453

Unlike the wrangle script, the first command will error when the input is
missing a "strain" column, but that is standard behavior across Augur and
other workflow tools, since we require "strain" to exist in the metadata.

Unlike the wrangle script which ensures that the "strain" column is the
first column, csvtk appends 'strain' as the last column. Since order of metadata
columns should not affect workflows, this is fine for the build. However, if
`augur.io.metadata.read_table_to_dict` (used by `augur curate`) will use the
first column as the ID column if unspecified, so the order matters if that
function is being called.

Comparing tsv files requires harmonizing their column names,
possibly with a csvtk cut command:

  FIELDS=`csvtk headers -t ORIGINAL.tsv | tr '\n' ','| sed 's/,$//'`
  csvtk cut -t -f $FIELDS NEW.tsv > NEW_harmonized.tsv
  diff ORIGINAL.tsv NEW_harmonized.tsv

Co-authored-by: John Huddleston <huddlej@gmail.com>
Co-authored-by: Victor Lin <13424970+victorlin@users.noreply.github.com>