notify-on-job-fail: Add job_name and repo_name args
Generalize the script by using new args to customize the Slack message
and point to the appropriate GitHub Action URL.

The repo_name arg was added based on diffs with subsequent copies of the
script that edited the Slack message:
- https://github.com/nextstrain/monkeypox/blob/5c461dc7e90cd70c1f16b193f82fd1666d4c95e2/ingest/bin/notify-on-job-fail
- https://github.com/nextstrain/forecasts-ncov/blob/70bf78f459a3706dd817ae5f711af3b74887d7b1/ingest/bin/notify-on-job-fail
- https://github.com/nextstrain/rsv/blob/ba171f4a43110382c38b6154be3febd50408d7bf/ingest/bin/notify-on-job-fail
- https://github.com/nextstrain/dengue/blob/247b2fd897361f2548627de1d97d45fae4115c5c/ingest/bin/notify-on-job-fail
- https://github.com/nextstrain/zika/blob/4ac8d526f9f14be10b7e8858ad469a40b72a505e/ingest/bin/notify-on-job-fail

Although the job_name arg is not necessary based on diffs with
subsequent copies of the script, it's nice to be able to customize the
Slack message for different ingest jobs. It's also a plus to mirror the
args for notify-on-job-start.

This will require all calls of the script to be update since job_name
and repo_name are required args.