Make the choice of "accession" on the config level
There was useful information on why "accession" is used over "strain" in
the docstring of update_example_data. Update that rule to use the
strain_id_field config, and move the context to every config file that
sets strain_id_field as "accession".

Don't set defaults when retrieving strain_id_field so "accession" is
only set on the config level.