Use pathogen-repo-build workflow for builds
Simplify build workflows by using the reusable Nextstrain
pathogen-repo-build workflow.

This does not update the fetch-and-ingest-branch workflow since that
will require additional changes to support the custom configs.

This does remove the `notify_pipeline_failed` step in the workflows.
They can be added as a separate job in the workflow if we think they
are still helpful.