Move common values to default config, allow partial overrides
Previously, either config_hmpxv1.yaml was used as the default as-is or
all required configuration were to be specified via Snakemake CLI
options. This resulted in much duplication of configuration across
different usages, and the default was never actually used in practice
except for in CI out of convenience of not specifying --configfile.

Switch to a two-tiered configuration setup with (1) a base "default"
tier of common values that can be applied to all workflow usages and (2)
a second tier of usage-specific config values which can also partially override

Since some required config entries (e.g. build_name) do not share common
values across the existing configs, this results in a workflow that is
only usable when additional configuration is defined. This will be
addressed by subsequent commits.