Re-organize sections
Rename "Installation" to "Software requirements" since the workflow is
not really something to be "installed". Move it to the top since
requirements naturally comes before usage.

Split "Data use" and "Update example data" sections to be their own
sections.

Move "Data use" to the bottom as is typical for acknowledgements.