wip: top level module
There are still many scripts left that would need to be prefixed with
the `workflow.current_basedir`.

I hadn't considered the fact that provided files via config params
would also need to be prefixed by the basedir variable.

With how unclear which files need to be prefixed and which do not, it
seems like more of pain than it's worth for using the Snakemake modules.