.github/workflows: Use centralized SLACK_CHANNELS variable
Using the centralized variable makes it easier for us to update the
slack channel value when needed.

As a part of this change, I've already added the SLACK_CHANNELS variable
to the repo. I opted to use the channel ID instead of the channel name
so that we have the option to rename the slack channel if desired.