ingest/upload: Refactor `files_to_upload` config to mapping
Simplify the `files_to_upload` config as a single mapping where the key
is the remote file name and the value is the local file instead of
maintaining a two lists of files. This ensures that we know exactly
which local file is uploaded to the remote file without worrying about
order or duplicates.