Fix ingest uploads
Fix bug that was introduced in https://github.com/nextstrain/mpox/pull/222

The testing done on branch runs do not include the Slack notifications,
so this bug was not revealed in the test run. I only noticed the
workflow as failing because our internal #monkeypox-updates channel had
been quiet for over a week.

This did not trigger our usual error notifications in Slack because
the error occurs during the DAG building process before the start
of the actual workflow run.