ingest: Always provide default config values
Part of work to update this repo to match the pathogen-repo-template.

Always provide default config values that can then easily be overridden
with --configfiles/--config options. This makes it simpler to change a
subset of config values or extend the default configs.

Renames the config/ to defaults/ to reflect this change.
Match changes in pathogen-repo-template in https://github.com/nextstrain/pathogen-repo-template/pull/27