phylo: move chores to build-configs
Part of work to update this repo to match the pathogen-repo-guide.

Since the chores rules are internal Nextstrain rules, they do not need
to be part of the core workflow. This also resolves the DAG confusion
that Snakemake occassionally runs into when running CI locally.¹

README.md includes the new instructions on how to invoke the workflow to
update the example data. This requires two config files:

1. The CI config to provide all required config params and to ensure
the example data uses correct `strain_id_field` for CI builds.
2. The chores config to include the custom rules

¹ https://github.com/nextstrain/mpox/issues/237