phylo: Move lapis to build-configs
Part of work to update this repo to match the pathogen-repo-guide.

LAPIS was the default data source until we switched to using
data.nextstrain.org.¹ This commit moves the LAPIS related rules to a
custom build-config and is able to successfully override the default
download/decompress rules. However, there are downstream changes to the
workflow that cause errors with the LAPIS data, e.g. hard-coded filters
on columns do not exist in the LAPIS data.²

Rather than spending the time to make sure the workflow runs with the
LAPIS data, I will remove the lapis build-configs in the next commit.
I wanted to do so separately to be able to easily revert the removal
if needed.

¹ https://github.com/nextstrain/mpox/commit/5d0791a06315eabb1eba43f6889996f0dcf824b2
² https://github.com/nextstrain/mpox/blob/34feb4a4e6f4d8b7929cc701734b15b8a5e6a6fc/phylogenetic/workflow/snakemake_rules/core.smk#L43