ingest: Remove `reverse` column from metadata
Previously this was filled in with empty strings by the custom
`ndjson-to-tsv-and-fasta` script. We've replaced vendored scripts
with `augur curate` commands, so we no longer need this script, but now
the workflow errors on this unexpected column.

As discussed in https://github.com/nextstrain/mpox/issues/209, this
column is not used anyways so just remove it.