Check for mutations and clades before removing
Fixes a bug in the fix_tree script when trying to remove mutations or
clades from lists where those values don't exist. The short-term fix is
to check for the presence of the values in the lists before trying to
remove them. The long term fix is to use a different tree builder than
eliminates the need for a "fix tree" script.