ingest: rename genbank.ndjson to ncbi.ndjson
Renaming to match other pathogens. Note this includes changing the file on S3 
to zstd compression as part of our effort switch to zstd.¹

Since this is a completely new file, I'm _not_ going to upload the xz file in 
parallel. This will require a manual download of genbank.ndjson.xz and re-upload
it as ncbi.ndjson.zst outside of this change so that our diff notification 
after the switch remains accurate.

¹ <https://github.com/nextstrain/mpox/issues/137>