ingest: simplify NCBI Datasets fields configs
Previously, I was using the ncbi-datset-field-map.tsv to rename NCBI Dataset 
fields to the fields we downloaded from NCBI Virus. Since we've renamed the file 
from genbank.ndjson to ncbi.ndjson in the previous commit, I figured we might as 
well stop renaming the fields in the NDJSON and use our standard config params.
The config changes are similar to the updates in the pathogen-repo-guide.¹

Important note that the final metadata.tsv columns do _not_ change with this 
change. I've kept them the same since outside users are using the metadata.tsv.

¹ <https://github.com/nextstrain/pathogen-repo-guide/commit/d9751bb1ed7478e4094289d57ea9ff163ee2673e>