ingest/nextstrain_automation: Copy `date_released` to `date_submitted`
Add custom rule to copy the `date_released` column to `date_submitted` so that 
outside users of our metadata.tsv have time to update their workflows to use 
the new `date_released` column. The plan is to remove the custom rule and the 
`date_submitted` column on 28 July 2025 so that we no longer have the 
duplicate column.

I'm adding this as a custom rule because this should only be needed in our 
automated runs for backwards compatibility of the public S3 files. Users who 
have updated their workflows and run ingest locally should not need the 
`date_submitted` column or can use custom configs to change the column name.