phylogenetic: Use list for config params that are multiple args
Allows Snakemake to do the right thing with quoted (:q) interpolations 
for config params that are passed in as multiple CLI args. Will automatically 
split string into list for backwards compatibility.

Based on similar changes done in zika
<https://github.com/nextstrain/zika/commit/11d2644441f4143db0bc560cf9b6b71ad53f4446>

Since the config function as_list are identical, this is definitely a candidate
for consolidating in nextstrain/shared.