phylogenetic: Use generic subsampling config param
Follows our guidelines on generic subsampling to use a single config param 
for the subsample rule.¹ I think this is an improvement over the previous 
subsample config params:
- it is no longer unclear which params require the CLI flag 
- removes the confusing dependency between `exclude_lineages` and `other_filters`

However, this does put the burden on the user to know  the augur filter flags 
and to format them correctly in the configs. I've created examples in the 
default build configs that uses the YAML folded style² to keep the args 
readable in the configs. This required an update to the yamlfmt config.

¹ <https://docs.nextstrain.org/en/latest/guides/bioinformatics/filtering-and-subsampling.html#generalizing-subsampling-in-a-workflow>
² <https://yaml.org/spec/1.2.2/#813-folded-style>

squash generic subsampling