phylogenetic: support disabling of subsampling groups via config
Previously, config overlays were unable to disable default subsampling defined 
in the `config["subsample"]` dict because Snakemake merges dicts in configs.
This change adds support for disabling default subsampling by ignoring subsample 
groups that have a null value. 

Users will still have to know the name of the default subsampling groups in 
order to nullify them in their config overlays. We should be doing a better 
job of surfacing these subsample names so that users don't have to read the 
source, but I'm not going to focus on that here as discussed in the PR
<https://github.com/nextstrain/mpox/pull/321#discussion_r2178746351>