🚧 Use augur subsample
The augur subsample command is built to replace the pattern of multiple
intermediate augur filter calls with simpler Snakemake rules and
configuration.

Note that this is a breaking change and the old configuration will no
longer work.

Due to limitations of augur subsample, I had to move the "include" file
from the top level of config to under a sample. For configs with
multiple samples, I chose the sample somewhat arbitrarily. We could
easily update augur subsample to accept the --include CLI option, but
I'm not sure it's necessary given this easy workaround.