phylogenetic: Add standardized multiple inputs support
Note this does _not_ remove the support for `private_metadata` and 
`private_sequences` config params because they are used by the private INRB
build to bypass the QC filtering for the contextual sequences.

In the future we could explore using the inputs `name` to change filtering 
params, but I'll punt on that for now.